Got a version rendering, but looking super funky. Struggling to make it look like water, seems like not having control over fresnel is a blocker here. On the plus side, using material property block which is nice - no more instances of ocean materials. And normals imported properly now and look legit (but only single scale and one non-animated layer so far).

+++ src/unity/Assets/Shaders/Ocean.shader
@@ -84,7 +84,7 @@ Shader "Ocean/Ocean"
 				// sample wave or terrain height, with smooth blend towards edges.
 				// would equally apply to heights instead of displacements.
 				// this could be optimized further.
-				void SampleDisplacements( in sampler2D i_dispSampler, in sampler2D i_oceanDepthSampler, in float2 i_centerPos, in float i_res, in float i_texelSize, in float i_geomSquareSize, in float2 i_samplePos, in float wt, inout float3 io_worldPos, inout float3 io_n, inout float io_determinant, inout half io_shorelineFoam )
+				void SampleDisplacements( in sampler2D i_dispSampler, in sampler2D i_oceanDepthSampler, in float2 i_centerPos, in float i_res, in float i_texelSize, in float i_geomSquareSize, in float2 i_samplePos, in float wt, inout float3 io_worldPos, inout float3 io_n, inout half io_determinant, inout half io_shorelineFoam )
 				{
 					if( wt < 0.001 )
 						return;
